backend/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
alembic==1.3.2
aniso8601==7.0.0
apparmor==2.13.3
appdirs==1.4.3
application-utility==1.3.1
arandr==0.1.10
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
autopep8==1.4.4
Babel==2.7.0
bauh==0.8.1
bcrypt==3.1.7
Beaker==1.11.0
boto3==1.10.34
botocore==1.13.34
btrfsutil==1.1.1
CacheControl==0.12.6
certifi==2019.9.11
cffi==1.13.2
channels==1.1.8
channels-api==0.4.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
commonmark==0.9.1
constantly==15.1.0
contextlib2==0.6.0
coverage==5.0.3
cryptography==2.8
cupshelpers==1.0
daphne==1.4.2
debtcollector==1.22.0
decorator==4.4.1
deluge==2.0.4.dev23
distlib==0.3.0
distro==1.4.0
Django==2.2
django-cors-headers==3.2.0
django-environ==0.4.5
django-filter==2.2.0
django-graphiql==0.4.4
django-graphql-jwt==0.2.1
djangorestframework==3.10.3
docopt==0.6.2
docutils==0.15.2
extras==1.0.0
fixtures==3.0.0
GeoIP==1.3.2
graphene==2.1.8
graphene-django==2.6.0
graphene-django-extras==0.4.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-ws==0.3.0
gufw==19.10.0
hiredis==1.0.1
html5lib==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.1.0
incremental==17.5.0
iso3166==1.0
iso8601==0.1.12
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lensfun==0.3.2
LibAppArmor==2.13.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
mapbox==0.18.0
MarkupSafe==1.1.1
more-itertools==8.0.0
msgpack==0.6.2
msgpack-python==0.5.6
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.6
netaddr==0.7.19
netifaces==0.10.9
npyscreen==4.10.5
oauth==1.0.1
olefile==0.45.1
ordered-set==3.1.1
oslo.config==6.12.0
oslo.context==2.23.0
oslo.db==5.0.2
oslo.i18n==3.25.0
oslo.log==3.45.0
oslo.serialization==2.29.2
oslo.utils==3.42.0
packaging==20.0
pacman-mirrors==4.16.2
paramiko==2.7.0
pbr==5.4.4
pep517==0.6.0
pep8==1.7.1
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
polyline==1.4.0
prettytable==0.7.2
progress==1.5
promise==2.2.1
protobuf==3.11.1
psutil==5.4.2
ptyprocess==0.6.0
purepng==0.2.0
pyaml==19.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.3
Pygments==2.5.1
PyGObject==3.34.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyRFC3339==1.0
pyrsistent==0.15.6
pysmbc==1.0.18
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python-libtorrent==1.2.3
python-mimeparse==1.6.0
python-subunit==1.3.0
python-xapp==1.8.1
python-xlib==0.26
pytoml==0.1.21
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
rally==2.1.0
ranger-fm==1.9.2
recommonmark==0.6.0
redis==2.10.6
rencode==1.0.6
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
retrying==1.3.3
rfc3986==1.3.2
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
Rx==1.6.0
s3transfer==0.2.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==2.0.0
speedtest-cli==2.1.2
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.11
sqlalchemy-migrate==0.13.0
sqlparse==0.3.0
stevedore==1.31.0
systemd-python==234
team==1.0
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.3.0
traceback2==1.4.0
Twisted==19.10.0
txaio==18.8.1
ueberzug==18.1.5
ufw==0.36
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.7.8
wadllib==1.3.2
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
